requirements.txt
folium==0.12.1.post1
geopandas==0.11.1
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.30
numpy==1.22.4
pandas==1.4.3
python_dateutil==2.8.2
requests==2.28.1
Shapely==1.7.1
streamlit==1.14.0
streamlit_folium==0.6.15
